+++ src/lib/api/providerTrainings.ts
@@ -1,13 +1,26 @@
 import type { RouterContext } from '@/routes/router'
-import { backendUrl, headersWithAuth } from './client'
+import { backendUrl, handleStatusCodes, headersWithAuth } from './client'
 import type { z } from 'zod'
-import type { ProviderTrainingUpdateRequestSchema } from '../schemas'
+import type {
+  ProviderTrainingUpdateRequestSchema,
+  ProviderTrainingResponseSchema,
+} from '../schemas'
 
-export async function getProviderTrainings(context: RouterContext) {
-  const res = await fetch(backendUrl('/provider/trainings'), {
+export async function getProviderTrainings(
+  context: RouterContext
+): Promise<z.infer<typeof ProviderTrainingResponseSchema>> {
+  const response = await fetch(backendUrl('/provider/trainings'), {
+    method: 'GET',
     headers: await headersWithAuth(context),
   })
-  return res
+
+  handleStatusCodes(context, response)
+
+  if (!response.ok) {
+    throw new Error(`Failed to fetch provider trainings: ${response.status}`)
+  }
+
+  return response.json()
 }
 
 export async function updateProviderTrainings(
